Improved deterministic random generation
In theory the previous method which could lead to short cycles since the
previous word was the only one which changed during a run and was used
as a seed. Now a running count of which word "we are at" is used instead
which should be better in this regard. And some evaluation here shows
that it works well.

I changed so that a user no longer has a global count of
"which attempted run is this". If a person has improved their bot and
wants to start from scratch they have to give it a new name.

Much simpler for everyone involved. So this number was before used in
the random generation. Now this number in next_word() instead is a count
of how many words have been guessed so far.